[Finishes #161307968] Remove typos and update relative imports
