+++ app/api/v1/models/users_model.py
@@ -1,0 +1,7 @@
+# class User:
+# 	def __init__(self,email,password):
+# 		self.email = email
+# 		self.password = password
+
+# class Attendant(User):
+# 	def __init__()
